requirements.txt
pymongo==4.10.1
flask==3.0.2
tensorflow==2.15.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
joblib==1.3.2
python-dotenv==1.0.1
matplotlib==3.8.3
APScheduler==3.10.4
gunicorn==21.2.0